+++ lib/packages/wikibase/data-model/RELEASE-NOTES.md
@@ -12,6 +12,7 @@
 #### Additions
 
 * Added `Statement::RANK_` enum
+* Added `Statement::addNewReference`
 
 #### Deprecations
 * Deprecated `Claim::RANK_` enum in favour of `Statement::RANK_` enum