+++ lib/packages/wikibase/data-model/RELEASE-NOTES.md
@@ -12,8 +12,10 @@
 #### Additions
 
 * Added `Statement::RANK_` enum
+* Added `Statement::addNewReference`
 
 #### Deprecations
+
 * Deprecated `Claim::RANK_` enum in favour of `Statement::RANK_` enum
 * Deprecated `Claim::getRank`