+++ lib/packages/wikibase/data-model/RELEASE-NOTES.md
@@ -8,6 +8,7 @@
 * Removed the class aliases for `ItemDiff` and `EntityDiff` (deprecated since 1.0)
 * Removed support for deserializing `EntityId` instances from version 0.4 or earlier
 * Removed `References` interface in favour of `ReferenceList`
+* The `Statement` constructor no longer supports a `Snak` parameter
 
 #### Additions